Made performance improvements
I added a class method that extracts the type for a query string. This
is a lot faster than parsing an entire query but has the disadvantage
that when you want to parse the query anyway for a certain filetype that
you would do double work then.

There is a configuration setting that allows the turning off of
calculating effective changes in SPARQL queries. This does mean that
when you send a query that has effective changes that they will not be
calculated (and thus the delta service will communicate no effective
changes for this query)