Changed the notification output format
The notification did not use application/json as the value of the
content-type header but instead application/x-www-form-urlencoded. I
switched the header to what it is supposed to send. It also set a
useless charset header to utf-8.

Both header settings have been commented out and I have added a todo to
remove them after investigation of why they may be needed.