Configuration file
I introduced a configuration file that will be read out using a
java.util.Properties class to manage.

The configuration settings that are currently in that file will probably
at some point still need to be expanded but are for now:
queryURL
updateURL

All others are not used at this point and will need to be implemented.