Added TODO to triple format
The triple format is not as it should be, I think this project should
move to a Triple class that consists of 3 Statements. The first two
would then by definition be URI Statements and the last one any
acceptable type.

This would make the entire process more robust and would solve all the
special cases that are in the product now thus making it much more maintainable.