Bugfix double quotes would not work
The query splitter used was not able to properly tokenize Strings that
contained escaped quotes since if a String contains "\"test\"" then the
value of the first character is '\"' as it should be, starting of the
String. The second character however would be '\\' and the third  '\"'
ending the String.

To solve this I have introduced a boolean to indicate if you are in a
String and the previous character was '\\' to write that character
whatever it is. So the aforementioned String gets properly parsed to a
single token.