Select block now supports multiple graphs
Before I assumed a SPARQL Query to only have 1 FROM statement but
apperently it can have many, ie:
SELECT *
FROM <GRAPH1>
FROM <GRAPH2>
WHERE
{
?s ?p ?o .
}

So now this is also supported. This does have complications for some of
the functionalities supported by the SPARQLStatement object. If you call
setGraph(graph) it will now remove ALL other graphs and set this block
to a new List<String> of graphs. After that the new graph is added to
that list. If you replace(oldGraph, new Graph) it first removes the old
graph from the list and then adds the new graph to the list.