Subscribers are now also configurable
The old system stays in place where you can subscribe to a running
system if you are interested in updates but now there is also the
possibility to set the subscribers in a config file;
subscribers.json. The format for this file is predefined and will be
read in with the Jackson library.

The file location needs to be set with an environment variable.